oh my god, wow https://www.youtube.com/watch?v=H46FRRvb-9o оно рабоает кароч

+++ src/main/java/fun/digitallpepper/logistic_service/controller/ConnectedNodesRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ConnectedNodesRequest {
     private String name1;
     private String name2;
-    private Long weight;
+    private Integer weight;
 }
